# ai-process-mining/docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "AI Process Mining Agent",
  "description": "Event logs \u2192 process discovery \u2192 metrics \u2192 bottlenecks \u2192 deviations \u2192 variants \u2192 AI explanation. The deterministic engine works without an LLM; the AI layer only explains computed metrics.",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/datasets/upload": {
   "post": {
    "tags": [
     "datasets"
    ],
    "summary": "Upload",
    "description": "Upload an event-log CSV (required columns: case_id, activity, timestamp).",
    "operationId": "upload_api_datasets_upload_post",
    "requestBody": {
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_upload_api_datasets_upload_post"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DatasetOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/datasets/{dataset_id}": {
   "get": {
    "tags": [
     "datasets"
    ],
    "summary": "Get Dataset",
    "operationId": "get_dataset_api_datasets__dataset_id__get",
    "parameters": [
     {
      "name": "dataset_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Dataset Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DatasetOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "datasets"
    ],
    "summary": "Delete Dataset",
    "operationId": "delete_dataset_api_datasets__dataset_id__delete",
    "parameters": [
     {
      "name": "dataset_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Dataset Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/datasets/{dataset_id}/validation": {
   "get": {
    "tags": [
     "datasets"
    ],
    "summary": "Validation",
    "operationId": "validation_api_datasets__dataset_id__validation_get",
    "parameters": [
     {
      "name": "dataset_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Dataset Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/datasets/{dataset_id}/validation/invalid-rows": {
   "get": {
    "tags": [
     "datasets"
    ],
    "summary": "Invalid Rows",
    "operationId": "invalid_rows_api_datasets__dataset_id__validation_invalid_rows_get",
    "parameters": [
     {
      "name": "dataset_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Dataset Id"
      }
     },
     {
      "name": "issue",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Issue"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 1000,
       "minimum": 1,
       "default": 100,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/datasets/{dataset_id}/validation/invalid-rows.csv": {
   "get": {
    "tags": [
     "datasets"
    ],
    "summary": "Invalid Rows Csv",
    "operationId": "invalid_rows_csv_api_datasets__dataset_id__validation_invalid_rows_csv_get",
    "parameters": [
     {
      "name": "dataset_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Dataset Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/datasets/{dataset_id}/cases/{case_id}": {
   "get": {
    "tags": [
     "datasets"
    ],
    "summary": "Case Detail",
    "operationId": "case_detail_api_datasets__dataset_id__cases__case_id__get",
    "parameters": [
     {
      "name": "dataset_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Dataset Id"
      }
     },
     {
      "name": "case_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Case Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/datasets/{dataset_id}/events": {
   "get": {
    "tags": [
     "datasets"
    ],
    "summary": "Events",
    "description": "Underlying events for drill-down (by activity, transition source\u2192target, or resource).",
    "operationId": "events_api_datasets__dataset_id__events_get",
    "parameters": [
     {
      "name": "dataset_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Dataset Id"
      }
     },
     {
      "name": "activity",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Activity"
      }
     },
     {
      "name": "source",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Source"
      }
     },
     {
      "name": "target",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Target"
      }
     },
     {
      "name": "resource",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Resource"
      }
     },
     {
      "name": "sort",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "pattern": "^(wait_desc|time_asc)$",
       "default": "wait_desc",
       "title": "Sort"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 1000,
       "minimum": 1,
       "default": 100,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "Body_upload_api_datasets_upload_post": {
    "properties": {
     "file": {
      "type": "string",
      "format": "binary",
      "title": "File"
     }
    },
    "type": "object",
    "required": [
     "file"
    ],
    "title": "Body_upload_api_datasets_upload_post"
   },
...
   "DatasetOut": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "error": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error"
     },
     "file_size_bytes": {
      "type": "integer",
      "title": "File Size Bytes"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "validation_report": {
      "anyOf": [
       {
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Validation Report"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "status",
     "file_size_bytes",
     "created_at"
    ],
    "title": "DatasetOut"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}